+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Remove warnings in PHP 8.2
+- Handle new API version accounting errors
 
 ## 0.6.0